Move map declarations to header
